add basic functionality, flesh out tests

+++ app/actions/issues/webhook.ts
@@ -1,4 +1,4 @@
-import { inject, ResponderParams } from 'denali';
+import { inject, Errors, ResponderParams } from 'denali';
 import ApplicationAction from '../application';
 
 interface WebhookPayload {
@@ -19,6 +19,7 @@ interface WebhookRepositoryPayload {
 export default class IssueWebhookAction extends ApplicationAction {
 
   githubApi = inject('service:github-api');
+  parser = inject('parser:raw');
 
   [key: string]: any;
 
@@ -38,22 +39,29 @@ export default class IssueWebhookAction extends ApplicationAction {
   }
 
   async reopened(body: WebhookPayload) {
-    this.opened(body);
+    return this.opened(body);
   }
 
   async closed({ issue, repository }: WebhookPayload) {
-    let record = await this.db.find('issue', `${ repository.full_name }#${ issue.id }`);
+    let record = await this.db.find('issue', issue.id);
+    if (!record) {
+      throw new Errors.NotFound();
+    }
     await record.delete();
   }
 
   async edited({ issue, repository }: WebhookPayload) {
-    let record = await this.db.find('issue', `${ repository.full_name }#${ issue.id }`);
+    let record = await this.db.find('issue', issue.id);
+    if (!record) {
+      throw new Errors.NotFound();
+    }
     Object.assign(record, this.issueDataFromPayload(issue));
     await record.save();
   }
 
   issueDataFromPayload(payload: WebhookIssuePayload) {
     return {
+      id: payload.id,
       title: payload.title
     };
   }
